third_party/blink/web_tests/external/wpt/css/css-logical/parsing/margin-block-inline-computed-expected.txt

This is a testharness.js-based test.
PASS Property margin-block-start value '10px' computes to '10px'
FAIL Property margin-block-end value '10%' computes to '20px' assert_equals: expected "20px" but got "0px"
PASS Property margin-inline-start value '30px' computes to '30px'
PASS Property margin-inline-end value '1em' computes to '40px'
FAIL Property margin-block-start value 'calc(10% + 40px)' computes to '60px' assert_equals: expected "60px" but got "0px"
PASS Property margin-block-end value 'calc(10px + 0.5em)' computes to '30px'
PASS Property margin-inline-start value 'calc(10px + 0.5em)' computes to '30px'
FAIL Property margin-inline-end value 'calc(10% + 40px)' computes to '60px' assert_equals: expected "60px" but got "40px"
PASS Property margin-block value '10px' computes to '10px'
PASS Property margin-block value '10px 20px' computes to '10px 20px'
PASS Property margin-inline value '30px' computes to '30px'
PASS Property margin-inline value '30px 40px' computes to '30px 40px'
Harness: the test ran to completion.
